+++ app/src/main/java/com/wafflestudio/snutt2/components/view/TextRect.kt
@@ -5,266 +5,133 @@ import android.graphics.Paint
 import android.graphics.Paint.FontMetricsInt
 import android.graphics.Rect
 
-/**
- * Created by makesource on 2016. 1. 24..
- */
 class TextRect(paint: Paint) {
-    // those members are stored per instance to minimize
-    // the number of allocations to avoid triggering the
-    // GC too much
-    private var metrics: FontMetricsInt? = null
-    private var paint: Paint? = null
-    private val starts = IntArray(MAX_LINES)
-    private val stops = IntArray(MAX_LINES)
+    private var metrics: FontMetricsInt
+    private var paint: Paint
+    private var lineheight = 0
+    private var maxWidth = 0
     private var lines = 0
-    private var textHeight = 0
+    private var maxlines = 0
+    private val starts = mutableListOf<Int>()
+    private val stops = mutableListOf<Int>()
+    private var textHeights = mutableListOf<Int>()
     private val bounds = Rect()
-    private var text: String? = null
+    private var text: String = ""
     private var wasCut = false
+    private var toDraw = true
 
-    /**
-     * Calculate height of text block and prepare to draw it.
-     *
-     * @param text - text to draw
-     * @param maxWidth - maximum width in pixels
-     * @param maxHeight - maximum height in pixels
-     * @returns height of text in pixels
-     */
-    fun prepare(text: String, maxWidth: Int, maxHeight: Int): Int {
-        lines = 0
-        textHeight = 0
+    fun prepare(text: String, maxWidth: Int, toDraw: Boolean) {
+        this.toDraw = toDraw
+        clear()
+        if (text.isEmpty()) this.toDraw = false
+        if (!toDraw) return
         this.text = text
-        wasCut = false
-
-        // get maximum number of characters in one line
-        paint!!.getTextBounds("i", 0, 1, bounds)
-        val maximumInLine = maxWidth / bounds.width()
-        val length = text.length
-        if (length > 0) {
-            val lineHeight = -metrics!!.ascent + metrics!!.descent + metrics!!.leading
-            var start = 0
-            var stop = if (maximumInLine > length) length else maximumInLine
-            // stop = min(length,maximuminline);
-            while (true) {
-                // skip LF and spaces
-                while (start < length) {
-                    val ch = text[start]
-                    if (ch != '\n' && ch != '\r' && ch != '\t' && ch != ' ') break
-                    ++start
-                }
-                var o = stop + 1
-                while (stop < o && stop > start) {
-                    o = stop
-                    var lowest = text.indexOf("\n", start)
-                    // 강의명만 생각
-                    paint!!.getTextBounds(text, start, stop, bounds)
-                    if (start <= lowest && lowest < stop || bounds.width() > maxWidth) // start ~ stop 까지 한 줄에 넣으려고 하는데 한 줄에 넘치면
-                        {
-                            --stop
-                            if (lowest < start || lowest > stop) {
-                                val blank = text.lastIndexOf(" ", stop)
-                                if (blank > start) lowest = blank
-                            }
-                            if (start <= lowest && lowest <= stop) {
-                                val ch = text[stop]
-                                if (ch != '\n' && ch != ' ') ++lowest
-                                stop = lowest
-                            }
-                            continue
-                        }
-                    break
-                }
-                if (start >= stop) break
-                var minus = 0
-
-                // cut off lf or space
-                if (stop < length) {
-                    val ch = text[stop - 1]
-                    if (ch == '\n' ||
-                        ch == ' '
-                    ) {
-                        minus = 1
-                    }
-                }
-                starts[lines] = start
-                stops[lines] = stop - minus
-                if (++lines > MAX_LINES) {
-                    wasCut = true
-                    break
-                }
-                if (textHeight + lineHeight > maxHeight) {
-                    wasCut = true
-                    break
-                }
-                textHeight += lineHeight
-                if (stop >= length) break
-                start = stop
-                stop = length
-            }
-        }
-        return textHeight
+        this.maxWidth = maxWidth
+        cutToLines()
+        maxlines = lines
     }
 
-    /**
-     * Calculate height of text block and prepare to draw it in a single line.
-     *
-     * @param text - text to draw
-     * @param maxWidth - maximum width in pixels
-     * @param maxHeight - maximum height in pixels
-     * @returns height of text in pixels
-     */
-    fun prepareSingleLine(text: String, maxWidth: Int, maxHeight: Int): Int {
-        paint!!.getTextBounds(text, 0, text.length, bounds)
-        if (bounds.width() == 0) return 0
-        if (bounds.width() <= maxWidth) {
-            val lineHeight = -metrics!!.ascent + metrics!!.descent + metrics!!.leading
-            if (lineHeight > maxHeight) {
-                wasCut = true
-                return 0
-            }
-            wasCut = false
-            lines = 1
-            starts[0] = 0
-            stops[0] = text.length
-            textHeight = lineHeight
-            return textHeight
+    fun getMaxLines() = maxlines
+
+    fun getTextHeight(reduceLine: Boolean = false): Int {
+        if (!toDraw || maxlines == 0) return 0 // 항목이 공백으로만 되어있는 예외적인 경우 처리
+        if (reduceLine) {
+            maxlines -= 1
+            wasCut = true
         }
+        return textHeights[maxlines - 1]
+    }
 
-        lines = 0
-        textHeight = 0
-        this.text = text
-        wasCut = true
+    fun getLeading(): Int = metrics.leading
 
-        // get maximum number of characters in one line
-        paint!!.getTextBounds("i", 0, 1, bounds)
-        val maximumInLine = maxWidth / bounds.width()
-        val length = text.length
-        if (length > 0) {
-            val lineHeight = -metrics!!.ascent + metrics!!.descent + metrics!!.leading
-            var start = 0
-            var stop = if (maximumInLine > length) length else maximumInLine
-            // stop = min(length,maximuminline);
+    fun draw(canvas: Canvas, left: Int, top: Int, bubbleWidth: Int, fgColor: Int) {
+        if (toDraw && maxlines > 0) {
+            attachEllipsis()
+            val before = -metrics.ascent
+            val after = metrics.descent + metrics.leading
+            var y = top
+            for (n in 0..<maxlines) {
+                y += before
+                val t = if (wasCut && n == maxlines - 1) {
+                    text.substring(starts[n], stops[n]) + "..."
+                } else {
+                    text.substring(starts[n], stops[n])
+                }
+
+                // 텍스트 가운데 정렬
+                var leftResult = left
+                paint.getTextBounds(t, 0, t.length, bounds)
+                val textWidth = bounds.right - bounds.left
+                leftResult += (bubbleWidth - textWidth) / 2
+                //
+                paint.color = fgColor
+                canvas.drawText(t, leftResult.toFloat(), y.toFloat(), paint)
+                y += after
+            }
+        }
+    }
+
+    private fun attachEllipsis() {
+        if (wasCut) {
             while (true) {
-                // skip LF and spaces
-                while (start < length) {
-                    val ch = text[start]
-                    if (ch != '\n' && ch != '\r' && ch != '\t' && ch != ' ') break
-                    ++start
-                }
-                var o = stop + 1
-                while (stop < o && stop > start) {
-                    o = stop
-                    var lowest = text.indexOf("\n", start)
-                    // 강의명만 생각
-                    paint!!.getTextBounds("$text...", start, stop + 3, bounds)
-                    if (start <= lowest && lowest < stop || bounds.width() > maxWidth) // start ~ stop 까지 한 줄에 넣으려고 하는데 한 줄에 넘치면
-                        {
-                            --stop
-                            if (lowest < start || lowest > stop) {
-                                val blank = text.lastIndexOf(" ", stop)
-                                if (blank > start) lowest = blank
-                            }
-                            if (start <= lowest && lowest <= stop) {
-                                val ch = text[stop]
-                                if (ch != '\n' && ch != ' ') ++lowest
-                                stop = lowest
-                            }
-                            continue
-                        }
-                    break
-                }
-                if (start >= stop) break
-                var minus = 0
+                val lastLineText = text.substring(starts[maxlines - 1]..<stops[maxlines - 1])
+                paint.getTextBounds("$lastLineText...", 0, lastLineText.length + 3, bounds)
+                if (bounds.width() <= maxWidth) break
+                stops[maxlines - 1] -= 1
+            }
+        }
+    }
 
-                // cut off lf or space
-                if (stop < length) {
-                    val ch = text[stop - 1]
-                    if (ch == '\n' ||
-                        ch == ' '
-                    ) {
-                        minus = 1
-                    }
+    private fun clear() {
+        lines = 0
+        maxlines = 0
+        starts.clear()
+        stops.clear()
+    }
+
+    private fun cutToLines() {
+        var textToPrepare = text
+        while (textToPrepare.isNotEmpty()) {
+            val textToCut = cutToSingleLine(textToPrepare)
+            if (textToCut.isEmpty()) break // 끝이 공백으로 끝나는 예외적인 경우 처리
+
+            starts.add(textToPrepare.indexOf(textToCut) + (if (lines == 0) 0 else stops[lines - 1]))
+            stops.add(starts[lines] + textToCut.length)
+            textHeights.add((if (lines == 0) 0 else textHeights[lines - 1]) + lineheight)
+            textToPrepare = text.substring(startIndex = stops[lines])
+            lines += 1
+        }
+    }
+
+    private fun cutToSingleLine(text: String): String {
+        var result = ""
+        for (i in 0..<text.length) {
+            val char = text[i]
+            if (char == '\n' && result.isNotEmpty()) return result
+            if (char == ' ' && result.isEmpty()) continue
+
+            result += char
+            paint.getTextBounds(result, 0, result.length, bounds)
+            if (bounds.width() > maxWidth) {
+                while (result.lastIndexOf(' ') in 0..<result.length - 1) {
+                    result = result.dropLast(1)
                 }
-                if (textHeight + lineHeight > maxHeight) {
-                    wasCut = true
-                    break
-                }
-                starts[lines] = start
-                stops[lines] = stop - minus + 3
-                if (++lines > MAX_LINES) {
-                    wasCut = true
-                    break
-                }
-                textHeight += lineHeight
-                if (stop >= length) break
+                result = result.dropLast(1)
                 break
             }
         }
-        return textHeight
+        return result
     }
 
-    /**
-     * Draw prepared text at given position.
-     *
-     * @param canvas - canvas to draw text into
-     * @param left - left corner
-     * @param top - top corner
-     */
-    fun draw(canvas: Canvas, left: Int, top: Int, bubbleWidth: Int, fgColor: Int) {
-        if (textHeight == 0) return
-        val before = -metrics!!.ascent
-        val after = metrics!!.descent + metrics!!.leading
-        var y = top
-        --lines
-        for (n in 0..lines) {
-            var t: String
-            y += before
-            t = if (wasCut && n == lines) {
-                text!!.substring(
-                    starts[n],
-                    if (stops[n] - 3 < starts[n]) starts[n] else stops[n] - 3,
-                ) + "..."
-            } else {
-                text!!.substring(starts[n], stops[n])
-            }
-
-            // 텍스트 가운데 정렬
-            var leftResult = left
-            paint!!.getTextBounds(t, 0, t.length, bounds)
-            val textWidth = bounds.right - bounds.left
-            leftResult += (bubbleWidth - textWidth) / 2
-            //
-            paint!!.color = fgColor
-            canvas.drawText(t, leftResult.toFloat(), y.toFloat(), paint!!)
-            y += after
-        }
-    }
-
-    /** Returns true if text was cut to fit into the maximum height  */
-    fun wasCut(): Boolean {
-        return wasCut
-    }
-
-    fun getLeading(): Int {
-        return metrics?.leading ?: 0
-    }
-
-    companion object {
-        // maximum number of lines; this is a fixed number in order
-        // to use a predefined array to avoid ArrayList (or something
-        // similar) because filling it does involve allocating memory
-        private const val MAX_LINES = 256
-    }
-
-    /**
-     * Create reusable text rectangle (use one instance per font).
-     *
-     * @param paint - paint specifying the font
-     */
     init {
         metrics = paint.fontMetricsInt
-        val temp = metrics.toString()
-        println("metrics$temp")
         this.paint = paint
+        this.lineheight = -metrics.ascent + metrics.descent + metrics.leading
     }
+}
+
+fun main() {
+    val a = "초급한문 1"
+    println(a.indexOf("초급한문"))
+    println(a.substring(0..<3))
 }
