+++ app/src/main/java/com/wafflestudio/snutt2/views/logged_in/home/timetable/TimetableCanvasObjects.kt
@@ -10,7 +10,6 @@ import androidx.compose.ui.unit.dp
 import androidx.core.content.res.ResourcesCompat
 import com.wafflestudio.snutt2.R
 import com.wafflestudio.snutt2.components.view.TextRect
-import com.wafflestudio.snutt2.components.view.TextRectTemp
 import com.wafflestudio.snutt2.lib.rx.dp
 import com.wafflestudio.snutt2.lib.rx.sp
 import com.wafflestudio.snutt2.ui.isDarkMode
@@ -58,7 +57,7 @@ object TimetableCanvasObjects {
 
     val lectureCellTextRect
         @Composable get() =
-            TextRectTemp(
+            TextRect(
                 Paint(Paint.ANTI_ALIAS_FLAG).apply {
                     textSize = 11f.sp(LocalContext.current)
                     typeface = ResourcesCompat.getFont(LocalContext.current, R.font.spoqa_han_sans_regular)
@@ -67,7 +66,7 @@ object TimetableCanvasObjects {
 
     val lectureCellTextRectReduced
         @Composable get() =
-            TextRectTemp(
+            TextRect(
                 Paint(Paint.ANTI_ALIAS_FLAG).apply {
                     textSize = 8.8f.sp(LocalContext.current)
                     typeface = ResourcesCompat.getFont(LocalContext.current, R.font.spoqa_han_sans_regular)
@@ -76,7 +75,7 @@ object TimetableCanvasObjects {
 
     val lectureCellSubTextRect
         @Composable get() =
-            TextRectTemp(
+            TextRect(
                 Paint(Paint.ANTI_ALIAS_FLAG).apply {
                     textSize = 12f.sp(LocalContext.current)
                     typeface = ResourcesCompat.getFont(LocalContext.current, R.font.spoqa_han_sans_regular)
@@ -85,7 +84,7 @@ object TimetableCanvasObjects {
 
     val lectureCellSubTextRectReduced
         @Composable get() =
-            TextRectTemp(
+            TextRect(
                 Paint(Paint.ANTI_ALIAS_FLAG).apply {
                     textSize = 9.6f.sp(LocalContext.current)
                     typeface = ResourcesCompat.getFont(LocalContext.current, R.font.spoqa_han_sans_regular)
@@ -94,7 +93,7 @@ object TimetableCanvasObjects {
 
     val lectureCellSubTextRectBold
         @Composable get() =
-            TextRectTemp(
+            TextRect(
                 Paint(Paint.ANTI_ALIAS_FLAG).apply {
                     textSize = 12f.sp(LocalContext.current)
                     typeface = ResourcesCompat.getFont(LocalContext.current, R.font.spoqa_han_sans_bold)
@@ -103,7 +102,7 @@ object TimetableCanvasObjects {
 
     val lectureCellSubTextRectReducedBold
         @Composable get() =
-            TextRectTemp(
+            TextRect(
                 Paint(Paint.ANTI_ALIAS_FLAG).apply {
                     textSize = 9.6f.sp(LocalContext.current)
                     typeface = ResourcesCompat.getFont(LocalContext.current, R.font.spoqa_han_sans_bold)
